+++ plugins/auth-backend-module-gs-provider/src/index.ts
@@ -1,0 +1,8 @@
+/***/
+/**
+ * The gs-providers backend module for the auth plugin.
+ *
+ * @packageDocumentation
+ */
+
+export { authModuleGsProviders as default } from './module';